chore (Validation and Feedback): Improving the validation and implementing feedback
Validate for url, white spaces, fixing  authentication issue and tests
[Delivers #163782662]
