+++ server/controllers/offices.js
@@ -1,11 +1,11 @@
 import Joi from 'joi';
-import PartiesModel from '../models/parties';
-import { validateParty } from './validationFunctions';
+import { validateOffice } from './validationFunctions';
+import OfficesModel from '../models/offices';
 
-class PartiesController {
-  // Get all parties
-  static getAllParties(req, res) {
-    PartiesModel.fetchAllParties(({ success, data }) => {
+class OfficesController {
+  // Get all offices
+  static getAllOffices(req, res) {
+    OfficesModel.fetchAllOffices(({ success, data }) => {
       // Check if the query was successful
       if (success) {
         return res.status(200).json({ status: 200, data });
@@ -17,17 +17,19 @@ class PartiesController {
     return null;
   }
 
-  static getSingleParty(req, res) {
-    PartiesModel.fetchPartyById(parseInt(req.params.id, 10), ({ success, data }) => {
+  static getSingleOffice(req, res) {
+    OfficesModel.fetchOfficeById(parseInt(req.params.id, 10), ({ success, data }) => {
       // Check if the query was successful
       if (success) {
-        // Check if the party exists
+        // Check if the office exists
 
         if (data.length === 0)
-          return res.status(404).json({ status: 404, error: 'The political party does not exist' });
+          return res
+            .status(404)
+            .json({ status: 404, error: 'The political office does not exist' });
 
-        // The party exists retrun to the user
-        return res.status(200).json({ status: 200, data: [data] });
+        // The office exists retrun to the user
+        return res.status(200).json({ status: 200, data });
       }
 
       // It is a server error
@@ -36,32 +38,31 @@ class PartiesController {
     return null;
   }
 
-  static createPoliticalParty(req, res) {
-    // Validate the party details
-    const { error } = validateParty(req.body);
+  static createPoliticalOffice(req, res) {
+    // Validate the office details
+    const { error } = validateOffice(req.body);
     if (error) return res.status(400).json({ status: 400, error: error.details[0].message });
-
-    // Check if the party exists before
-    PartiesModel.fetchPartyByName(req.body.name, ({ success, data }) => {
+    // Check if the office exists before
+    OfficesModel.fetchOfficeByNameAndType(req.body.name, req.body.type, ({ success, data }) => {
       if (!success) {
         // It is a server error
         return res.status(500).json({ status: 500, error: data });
       }
 
-      // Check if the party exists
+      // Check if the office exists
       if (data.length !== 0)
-        // The party already exists
-        return res.status(409).json({ status: 409, error: 'This political party already exists' });
+        // The office already exists
+        return res.status(409).json({ status: 409, error: 'This political office already exists' });
 
-      // Create the new party
-      const newParty = {
+      // Create the new office
+      const newOffice = {
         name: req.body.name,
-        hqAddress: req.body.hqAddress,
+        type: req.body.type,
         logoUrl: req.body.logoUrl
       };
 
-      //   Insert the new party in the array
-      PartiesModel.createNewParty(newParty, ({ successs, dataa }) => {
+      //   Insert the new office in the array
+      OfficesModel.createNewOffice(newOffice, ({ successs, dataa }) => {
         if (successs) {
           // Return the new user details to the user
           return res.status(201).json({ status: 201, data: dataa });
@@ -73,21 +74,21 @@ class PartiesController {
     return null;
   }
 
-  static deletePoliticalParty(req, res) {
-    // Check if the party exists
-    PartiesModel.fetchPartyById(parseInt(req.params.id, 10), ({ success, data }) => {
+  static deletePoliticalOffice(req, res) {
+    // Check if the office exists
+    OfficesModel.fetchOfficeById(parseInt(req.params.id, 10), ({ success, data }) => {
       if (!success) {
         // It is a server error
         return res.status(500).json({ status: 500, error: data });
       }
 
-      // Check if the party exists
+      // Check if the office exists
       if (data.length === 0) {
-        return res.status(404).json({ status: 404, error: 'Party does not exist' });
+        return res.status(404).json({ status: 404, error: 'Office does not exist' });
       }
 
       // Delete from server
-      PartiesModel.deletePartyById(parseInt(req.params.id, 10), ({ successs }) => {
+      OfficesModel.deleteOfficeById(parseInt(req.params.id, 10), ({ successs }) => {
         if (!successs) {
           return res.status(500).json({ status: 500, error: data });
         }
@@ -99,31 +100,29 @@ class PartiesController {
     return null;
   }
 
-  static editParticularPoliticalParty(req, res) {
+  static editParticularPoliticalOffice(req, res) {
     // Validate the name
     const schema = {
       name: Joi.string()
         .min(10)
         .max(40)
         .required()
-        .trim()
-        .strict()
     };
     const { error } = Joi.validate({ name: req.params.name }, schema);
     if (error) return res.status(400).json({ status: 400, error: error.details[0].message });
 
-    PartiesModel.fetchPartyById(parseInt(req.params.id, 10), ({ success, data }) => {
+    OfficesModel.fetchOfficeById(parseInt(req.params.id, 10), ({ success, data }) => {
       if (!success) {
         return res.status(500).json({ status: 500, error: data });
       }
 
-      // Check if there is party
+      // Check if there is office
       if (data.length === 0) {
-        return res.status(404).json({ status: 404, error: 'Party does not exist' });
+        return res.status(404).json({ status: 404, error: 'Office does not exist' });
       }
 
-      // Update the party
-      PartiesModel.updatePartyNameById(req.params.id, req.params.name, ({ successs, dataa }) => {
+      // Update the office
+      OfficesModel.updateOfficeNameById(req.params.id, req.params.name, ({ successs, dataa }) => {
         if (successs) {
           // Return the new user details to the user
           return res.status(200).json({ status: 200, data: dataa });
@@ -136,4 +135,4 @@ class PartiesController {
   }
 }
 
-export default PartiesController;
+export default OfficesController;
