+++ server/controllers/parties.js
@@ -1,11 +1,11 @@
 import Joi from 'joi';
-import { validateOffice } from './validationFunctions';
-import OfficesModel from '../models/offices';
+import PartiesModel from '../models/parties';
+import { validateParty } from './validationFunctions';
 
-class OfficesController {
-  // Get all offices
-  static getAllOffices(req, res) {
-    OfficesModel.fetchAllOffices(({ success, data }) => {
+class PartiesController {
+  // Get all parties
+  static getAllParties(req, res) {
+    PartiesModel.fetchAllParties(({ success, data }) => {
       // Check if the query was successful
       if (success) {
         return res.status(200).json({ status: 200, data });
@@ -17,19 +17,17 @@ class OfficesController {
     return null;
   }
 
-  static getSingleOffice(req, res) {
-    OfficesModel.fetchOfficeById(parseInt(req.params.id, 10), ({ success, data }) => {
+  static getSingleParty(req, res) {
+    PartiesModel.fetchPartyById(parseInt(req.params.id, 10), ({ success, data }) => {
       // Check if the query was successful
       if (success) {
-        // Check if the office exists
+        // Check if the party exists
 
         if (data.length === 0)
-          return res
-            .status(404)
-            .json({ status: 404, error: 'The political office does not exist' });
+          return res.status(404).json({ status: 404, error: 'The political party does not exist' });
 
-        // The office exists retrun to the user
-        return res.status(200).json({ status: 200, data });
+        // The party exists retrun to the user
+        return res.status(200).json({ status: 200, data: [data] });
       }
 
       // It is a server error
@@ -38,32 +36,32 @@ class OfficesController {
     return null;
   }
 
-  static createPoliticalOffice(req, res) {
-    // Validate the office details
-    const { error } = validateOffice(req.body);
+  static createPoliticalParty(req, res) {
+    // Validate the party details
+    const { error } = validateParty(req.body);
     if (error) return res.status(400).json({ status: 400, error: error.details[0].message });
 
-    // Check if the office exists before
-    OfficesModel.fetchOfficeByNameAndType(req.body.name, req.body.type, ({ success, data }) => {
+    // Check if the party exists before
+    PartiesModel.fetchPartyByName(req.body.name, ({ success, data }) => {
       if (!success) {
         // It is a server error
         return res.status(500).json({ status: 500, error: data });
       }
 
-      // Check if the office exists
+      // Check if the party exists
       if (data.length !== 0)
-        // The office already exists
-        return res.status(409).json({ status: 409, error: 'This political office already exists' });
+        // The party already exists
+        return res.status(409).json({ status: 409, error: 'This political party already exists' });
 
-      // Create the new office
-      const newOffice = {
+      // Create the new party
+      const newParty = {
         name: req.body.name,
-        type: req.body.type,
+        hqAddress: req.body.hqAddress,
         logoUrl: req.body.logoUrl
       };
 
-      //   Insert the new office in the array
-      OfficesModel.createNewOffice(newOffice, ({ successs, dataa }) => {
+      //   Insert the new party in the array
+      PartiesModel.createNewParty(newParty, ({ successs, dataa }) => {
         if (successs) {
           // Return the new user details to the user
           return res.status(201).json({ status: 201, data: dataa });
@@ -75,21 +73,21 @@ class OfficesController {
     return null;
   }
 
-  static deletePoliticalOffice(req, res) {
-    // Check if the office exists
-    OfficesModel.fetchOfficeById(parseInt(req.params.id, 10), ({ success, data }) => {
+  static deletePoliticalParty(req, res) {
+    // Check if the party exists
+    PartiesModel.fetchPartyById(parseInt(req.params.id, 10), ({ success, data }) => {
       if (!success) {
         // It is a server error
         return res.status(500).json({ status: 500, error: data });
       }
 
-      // Check if the office exists
+      // Check if the party exists
       if (data.length === 0) {
-        return res.status(404).json({ status: 404, error: 'Office does not exist' });
+        return res.status(404).json({ status: 404, error: 'Party does not exist' });
       }
 
       // Delete from server
-      OfficesModel.deleteOfficeById(parseInt(req.params.id, 10), ({ successs }) => {
+      PartiesModel.deletePartyById(parseInt(req.params.id, 10), ({ successs }) => {
         if (!successs) {
           return res.status(500).json({ status: 500, error: data });
         }
@@ -101,31 +99,29 @@ class OfficesController {
     return null;
   }
 
-  static editParticularPoliticalOffice(req, res) {
+  static editParticularPoliticalParty(req, res) {
     // Validate the name
     const schema = {
       name: Joi.string()
         .min(10)
         .max(40)
         .required()
-        .trim()
-        .strict()
     };
     const { error } = Joi.validate({ name: req.params.name }, schema);
     if (error) return res.status(400).json({ status: 400, error: error.details[0].message });
 
-    OfficesModel.fetchOfficeById(parseInt(req.params.id, 10), ({ success, data }) => {
+    PartiesModel.fetchPartyById(parseInt(req.params.id, 10), ({ success, data }) => {
       if (!success) {
         return res.status(500).json({ status: 500, error: data });
       }
 
-      // Check if there is office
+      // Check if there is party
       if (data.length === 0) {
-        return res.status(404).json({ status: 404, error: 'Office does not exist' });
+        return res.status(404).json({ status: 404, error: 'Party does not exist' });
       }
 
-      // Update the office
-      OfficesModel.updateOfficeNameById(req.params.id, req.params.name, ({ successs, dataa }) => {
+      // Update the party
+      PartiesModel.updatePartyNameById(req.params.id, req.params.name, ({ successs, dataa }) => {
         if (successs) {
           // Return the new user details to the user
           return res.status(200).json({ status: 200, data: dataa });
@@ -138,4 +134,4 @@ class OfficesController {
   }
 }
 
-export default OfficesController;
+export default PartiesController;
