+++ server/controllers/votes.js
