+++ server/models/votes.js
