+++ server/routes/votes.js
